Fixed Errors where I tried to use a Long type instead of int for Id

+++ src/main/java/com/finance/repositories/FundRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.finance.models.Fund;
 
 @Repository
-public interface FundRepository extends CrudRepository<Fund, Long> {
+public interface FundRepository extends CrudRepository<Fund, Integer> {
 
 }
